[iOS] Restore the correct identity when canceling sign-in
This patch fixes the following scenario:
 * Sign-in (without sync) with identity 1
 * Open settings, and tap on Turn on sync
 * Select identity 2
 * Tap on "settings" link
 * Cancel the turn on sync dialog

When tapping on settings links, identity 1 is signed out, and identity 2
is signed in, to be able to open advanced settings.
The bug is that UserSigninMediator doesn't revert identity 1 when
canceling the dialog.

This patch saves the identity when UserSigninCoordinator is started
(at the same time than the sign-in state). And UsersSigninMediator
tests to make sure the current signed in identity is identical to
the saved identity. If not, the mediator does a sign-out and signs in
to the saved identity.

Fixed: 1268850
Change-Id: Iab45da725ef9a0cfde9be412646cb721cdc7bef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3288683
Commit-Queue: Jérôme <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942557}
NOKEYCHECK=True
GitOrigin-RevId: a83ce7a96f3ac0d2dba616b91a3e76f993ccab60