Factor BVC's web state delegate installation into a mediator
Currently the BVC is responsible for setting the delegates of several
tab helpers with UI dependencies. This requires the BVC to observe the
web state list to update these helpers when new tabs are added, which
is outside of a view controller's responsibilities.

The BVC also implementing these delegate protocols is also a problem,
but not one addressed by this CL.

This CL factors this job into a mediator (TabLifecycleMediator) which
the BrowserCoordinator owns. It handles injecting UI dependencies into
new web states (and removing them from departing ones). To do this,
the mediator is provided with a slew of objects (factories and other
objects) as well as a catchall delegate which conforms to all of the
tab helper delegate protocols that the BVC implements.

To facilitate this, this CL adds a category on the BVC which includes
its conformance to those protocols, as well as a few other properties
which are needed as dependencies. This category is only used by the
BVC and BrowserCoordinator.

A few tab helpers that were being configured in the BVC didn't need to
be, and are moved into AttachTabHelpers().

(BrowserCoordinator does similar delegate attachment and unattachment
via WebStateList observation; a follow-up CL also moves this into the
TabLifecycleMediator).

Change-Id: I09d354a0c4a4172cf4d8a2d9043f7d328451f89a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268917
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942613}
NOKEYCHECK=True
GitOrigin-RevId: d3e95ce49a672622c4053165e99d9954a5167704