[iOS] Re-enable disabled NTP tests
testAccessibility: Was caused by a missing label in Mulder. Seems to be
fixed now.

testPositionRestoredWithOmniboxFocused: NTP improvements since August
seem to have fixed the test on iPads.

testToggleFeedEnabled: The flakes happened when the header was not
visible but a reference to it in the tests caused it to be not nil.
Changed condition to check visibility instead of nil-checking.

Bug: 1237925, 1194106, 1243222
Change-Id: I4f17548e4538e8b9318ca5de73ee4d7f456e604f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3290003
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#942736}
NOKEYCHECK=True
GitOrigin-RevId: 715f359e957d91ff7527f69d36ae5e866c088f6f