[ios] Cleanup WebStateImpl API
Change member field declaration to provide a default value in
order to ensure they are correctly initialized by all the
constructors.

Simplify the definition of the callback of RunJavaScriptDialog
by using an helper function.

Remove superfluous comments and DCHECK about ObserverList (the
base class already check that an observer is not added twice
nor removed when it is not present).

Change some getter to be marked as `const`.

Bug: 1264453
Change-Id: I5d353a449f59843f5f9d3de8994a513f100aa7fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3285403
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943108}
NOKEYCHECK=True
GitOrigin-RevId: e7af9f49393b001a2b636a44d35a99ae78a061bc