[iOS]Hardcode navigation bar style for <= iOS 14 to workaround iOS issue
An issue with UINavigationController is observed for iOS versions 14 and
under. If the navigation bar is set to be transparent on scroll edge and
when a new view is pushed upon the navigation controller, the navigation
bar would not be resized to fit the navigation items in the new view.

In this case, when the user taps "< back" from Passwords in Other Apps
view controller, the search bar in the Passwords table view would not be
included in the navigation bar. In dark mode, the search bar has a
different background color from the navigation bar buttons, and in light
mode, a horizontal separator would appear above the search bar, leading
to confusing user experience.

To workaround this bug, this CL hardcodes the style of the entire
settings navigation controller to have a non-transparent bar, and child
views would turn on/off transparency only when needed.

Note that since this issue is resolved by Apple in iOS 15, the predicate
"!base::ios::IsRunningOnIOS15OrLater()" is used to wrap the temporary
fix. These blocks should be removed once iOS 14 is no longer supported.

Bug: 1268684
Change-Id: I4a2624b527422d4604853d9b72f13e6081d82139
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3283345
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943129}
NOKEYCHECK=True
GitOrigin-RevId: a7c61dfe924e16fa9ed9d77de74ef7ce78bd9fcb