[iOS] Move tab helper delegate setup out of BrowserCoordinator
This CL gets BrowserCoordinator out of the business of setting up
tab helper delegates. This also removes the need for BrowserCoordinator
to be a WebStateListObserver.

All of the tab helper delegate setup is moved into the tab lifecycle
mediator. The OpenInCoordinator creation and setup was also handled
as part of delegate installation, but this doesn't seem to be necessary,
so it's moved to align with other coordinator creation and shutdown.

In addition to adding the setup of the new delegates to
TabLifecycleMediator, the install/uninstall code is reorganized to
more clearly group different types of tab helper creation and/or
attachment, and to remove spurious nil checks.

Bug: 906525
Change-Id: Ie5fdd31c183fb975e2909236ec7f4c6f4ce9916f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3270675
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943232}
NOKEYCHECK=True
GitOrigin-RevId: f91729ab7ae1dab870d7b3aa93207ff6b60be747