[ios] Further cleanup of WebStateImpl interface
Group the methods declaration so that overridden methods are
not in the middle of WebStateImpl's specific methods.

Sort the methods so that they are in the same order in the
declaration and definition, and so that overridden methods
are in the same order as in the super class.

Mark most of the overridden methods as `final`.

Fix #include (and use #import as recommended by the style guide)
and remove unnecessary `web::` prefixes.

Bug: 1264453
Change-Id: I1d355110045e64ef5f15f49807971c4172101adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3289130
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943456}
NOKEYCHECK=True
GitOrigin-RevId: 3ebc0884f9a6ad09715a4f3be424359b9eb9b0b3