[iOS][cpe-promo] Utilize provider for open settings action
PasswordsInOtherAppsMediator should use provider method to react to the
button to open settings so that the behavior could be controlled by
Chrome on a higher level.

Also, to make sure the provider and delegate method always works, this
CL sets the flag "enable-shortened-password-auto-fill-instruction" to
be always on, so the flag is added to the "never-expire" list.

Bug: 1269843
Change-Id: I6d56ea47774879fd290677a682080c84d3a746c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3289829
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943577}
NOKEYCHECK=True
GitOrigin-RevId: 9a91e99191148042f3951390346efd49fe338e5e