Stop Price Chip from falling off Tab Grid Cell.
In the event of large font sizes or very large prices, the chip
which displays price drops falls off the Grid Cell. This CL
adds a constraint to stop this and adds in some spacing at the
end (as per UX requirements).

Bug: 1272072
Change-Id: Ia754776fc4a631ab903ead559142b9bd5e948fce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3293746
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#943711}
NOKEYCHECK=True
GitOrigin-RevId: aad018694e7d5d6f8beba57d5f62a25cb5ee04fd