[iOS] Stop promo text from disappearing in landscape mode
The root cause here is the aspect ratio constraint on the image view.
When the image view is hidden, the stack view sets the height of the
image view to 0. Because of the aspect ratio constraint, this also
forces the width of the image view to 0. And then (I think) because
stack view's alignment is "Fill", there are constraints between the
width of the stack view and the width of each element, so when the
image's width is 0, that forces the stack view's width to be 0 too.

This CL fixes the issue by turning on and off the constraint when the
image view is hidden.

Bug: 1272824
Change-Id: Ieab6fb46c4d40f32f0de3a89b64279452713ee20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297918
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#945397}
NOKEYCHECK=True
GitOrigin-RevId: c6dd50ef42086d1e2e803875f07b17902380839d