[ios] Support "unrealized" state for WebStateImpl
Add RealizedWebState and SerializedData as two inner classes of
WebStateImpl (defining them as inner class allow them to peak into
the WebStateImpl for easy access to the ObserverLists). Those two
classes corresponds respectively to the realized and unrealized
state for WebStateImpl.

Change the implementation of WebStateImpl to move all the logic
into either RealizedWebState or SerializedData (when possible).
WebStateImpl manages the transition from SerializedData to
RealizedWebState when necessary (view an helper function).

One important invariant is that at all point in time exactly one
of `saved_` or `pimpl_` is non-null. This invariant is maintained
at all time for code outside of WebStateImpl (the only point when
it is not maintained should not be observable from the current
thread). Since WebState is not thread-safe, it is okay to break
the invariant as long as it cannot be observed externally.

Change the WebStateImpl API to not exposes methods/inheritances
that are implementation detail (and that are now implemented by
RealizedWebState).

Fix the following tests that are broken by the changes as the
class WebStateImpl don't implement NavigationManagerDelegate:

-   SessionStorageBuilderTest.*: allow injecting a fake
    CRWWebViewNavigationProxy into WebStateImpl when it
    is constructed and remove WebStateImpl sub-class as
    it is no longer needed

-   CRWWebControllerTest.TestPendingItem: do not check
    the state of the NavigationManagerDelegate since it
    is now an implementation detail

Mark WebStateImpl as final now that WebStateWithMockProxy has been
deleted (since it was the only sub-class of WebStateImpl).

Bug: 1264453
Change-Id: I8e5c6b31affcb1619dd87093a34e03b82d7c9062
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3285464
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945613}
NOKEYCHECK=True
GitOrigin-RevId: 4c7e8b2c456604f4142af3c1450c4d541129f9ce