Component Updater: Refactor disable-updates policy.
The handling of the policy and combination with component data is moved
out of update_client (which is used by many embedders) and into
component_updater.

To achieve this, component_updater's Register no longer takes a strict
subset of CrxComponent information as an argument. A new type
(ComponentRegistration) is introduced to hold a component's registration
information with the component updater, and it is used to construct
CrxComponents on the fly in the crx data callback. The policy is applied
at that time.

There should be no change in behavior, but the simplication of
update_client's API makes the behavior of chrome/updater more clear.

Bug: 1094024
Change-Id: I0c1992acd3baf4ef30adb5ce5d191ff13df9c3f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291368
Auto-Submit: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Hazem Ashmawy <hazems@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#946615}
NOKEYCHECK=True
GitOrigin-RevId: 4df664317fc9f29f608ea8405a55b26c8bd1d8ca