Enable the HeapProfilerReporting feature by default
Adds probability parameters that are checked on the client side, so only
a fraction of clients with the feature enabled will actually run the
profiler and upload reports. This way the variations server can stop
sending any state, and the default settings will match the existing
server-side setup.

The variations server can go back to controlling which population runs
the heap sampler by disabling HeapProfilerReporting for users that
should not run it, and enabling it while setting the probability
parameters to 1.0 for users that should run it.

Bug: 1271555
Change-Id: Ib07ff14e2a34a894a4e05962327ddcb45cf71acb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3292515
Auto-Submit: Joe Mason <joenotcharles@google.com>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947768}
NOKEYCHECK=True
GitOrigin-RevId: 3457761da959886bdf35bc9bfa40e0622450da5d