[ios] Support unrealized WebStates in GetTabTitle()
Change GetTabTitle() prototype to take a const pointer to a
WebState (as it does not need to modify it). This means that
the implementation will call the const version of the method
GetNavigationManager() that returns null if the WebState is
unrealized.

Extract the code to check whether the WebState is presenting
a download to a helper method (that needs to deal with a null
NavigationManager).

Bug: 1264453
Change-Id: Id1a7bac8028ef829ad2a7cefaad1c900494c325c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3313123
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#947983}
NOKEYCHECK=True
GitOrigin-RevId: 7d7b64e6fcd8eaab63d818518a958f5c0e7ce3c8