[ios] Cleanup CRWSessionStorage and CRWNavigationItemStorage
Use lightweight generics annotation on itemStorages property
of CRWSessionStorage (to indicate the type of the items that
are stored in that array) and propagate the annotation in the
client code.

Change CRWNavigationItemStorage's URL, virtualURL and title
properties to pass the take and return the values by const
reference (this will be useful to implement retrieving the
corresponding values for "unrealized" WebStates in followup
change list).

Add some missing static_cast<> to convert between NSUInteger
and int, and add an explicit comparison with 0 to convert a
NSUInteger to a boolean.

Bug: 1264453
Change-Id: I428062a48b241c6c545453ba2fcfa8e47e758c96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3311073
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948379}
NOKEYCHECK=True
GitOrigin-RevId: 8d342484c6b4b842e08a0c0d1d319d6e4a07148a