[iOS][test runner] Do not report BUILD_INTERRUPTED to indicate crash
Historically iOS runner has been reporting BUILD_INTERRUPTED as a dummy
test result, when there are test suite crashes making the runner unsure
if all tests that should run have a result. However, it's not a standard
way to report crash, and it's confusing flakiness automation tools as
they consider it as a test name. This CL removes it as a dummy test
result and update test runner logs to report this. There is also a
signal of crash in "interrupted" field in JSON result format.

Bug: 1271964
Change-Id: Id7d28e21364e56c292272c83cd9bc708b99446ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3310200
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948637}
NOKEYCHECK=True
GitOrigin-RevId: 9c21a983cc9e5b756bfe2188558b5ffd88ab6069