[IOS] Fix New Search when Discover is hidden.
Fixes the issue where focusing the fakebox on the NTP when the Discover
feed is hidden doesn't show the omnibox (this happens when using the
'New Search' option from the new tab menu).

The fix is to change how the minimum NTP height is calculated -- this is
the 2-line change to ContentsSuggestionsLayout.

The remainder of the CL is adding EG tests for the New Search menu
item, with and without enabling the Discover feed. This also adds
UI-driven support for showing and hiding the feed.

Bug: 1273046, 1194106
Change-Id: Ie7d54cf65841644e0b2605d0436925fc618a771d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3314898
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#948974}
NOKEYCHECK=True
GitOrigin-RevId: 10a91ad84fb58202c3a0685d9a7b1b7a69f79696