Address feedback on WebStateDependencyInstallationObserver
This CL modifies DependencyInstallerBridge to be a subclass of
WebStateDependencyInstallationObserver, so as to reduce the number of
member variables users of that class have to add. This was requested
after landing the original class.

It also corrects a minor style guide violation (using snake_case in an
Objective-C protocol) and adds missing comments.

Bug: 1277163
Change-Id: I68f6fcb0f38739b7119be996ae4aa19d6ddaa84f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3318373
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#949055}
NOKEYCHECK=True
GitOrigin-RevId: 89b57de8e59afe60a89e0fe165631cad513dcb6f