[iOS][cpe-promo] End-to-end test setup for passwords in other apps
This CL setups end to end tests (EarlGrey) for Passwords In Other Apps
view controller. It creates the mock for PasswordAutoFillStatusManager
as well as an app interface to interact with it.

It also contains a first test case to make sure the app interface and
the mock manager works well. Actual test cases will come in a subsequent
CL.

Bug: 1271616
Change-Id: I5b795c123bf2952fcd736ec7f123103257763ba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3307613
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950082}
NOKEYCHECK=True
GitOrigin-RevId: e2def3c3190af752e8097453565e76854c754f14