[ios] Fix deletion of NSTemporaryDirectory()
Upon cancelling the download, DownloadNativeTaskBridge needs to
tell WkDownload to start the download with a path (otherwise an
exception is raised by WKDownload).

To avoid cluttering the user disk, the code schedule a task to
delete the file soon after the cancel has happened. There was
a bug as `self.suggestedFilename` could be empty. In that case
the download path was NSTemporaryDirectory() itself which was
then deleted, causing all access to the directory to fail after
that point.

This caused DownloadSessionTaskImplTest.FileDeletion to fail if
run as part of all tests, but to pass individually. Since the
bot retry failing tests, this was not found on the CQ.

The bug is serious as it can happen in production, and then all
creation of temporary files will fail. Force the use of a non
empty filename in `-[DownloadNativeTaskBridge cancel]` by using
a UUID (guaranteed to be non-emtpy). Since this is only used if
the download is cancelled after it has really started, the path
is not visible from other part of the code, so a random name is
not a problem.

Bug: none
Change-Id: Icf0c70a211c13e8514d2e365068878c5f166abe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3326950
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#950215}
NOKEYCHECK=True
GitOrigin-RevId: 5808f785d6244d5c4eb9c6f6ac9c7f170ef5c117