[ios] Manage stable identifier in WebState
The //ios/chrome code needs an identifier that is stable across
application restart to identify uniquely a WebState (e.g. it is
used to create the name of the snapshot).

This identifier was managed by TabIdTabHelper which is complex.
Move it to WebState directly which allow to simplify how it is
saved or loaded.

This will allow using the identifier in the serialisation code.

To avoid this CL from getting too large, the class TabIdTabHelper
is not removed and instead is converted to return the value from
WebState::GetStableIdentifier().

Bug: 1276776
Change-Id: I1fba68aa49406c0aa2da0d825caed53e79442f8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3324543
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951447}
NOKEYCHECK=True
GitOrigin-RevId: 60cd53c5f4d7a6f482a1d8b6314a4c353444e737