Refactor ReportRequest definition to use absl::variant
The ReportRequest is a type alias based on Operating system. It's
enterprise_management::ChromeOsUserReportRequest for Chrome OS and
enterprise_management::ChromeDesktopReportRequest for other platforms.

However, this is no longer work when adding the third proto which is
available on all platforms.

This CL create a new ReportRequest class to wrap the proto and use
absl::variant to store the request proto. This class is used to pass
request instance between report generator, uploader and scheduler.

Bug: 1261945
Change-Id: I0944fc2b82489523f3a1a6030f0c1daa6feab5b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3324322
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#951649}
NOKEYCHECK=True
GitOrigin-RevId: 6af348ddff8c41115dfc1c87a86ef13ee8e7d2c6