[iOS] Allow new overflow menu destinations to wrap to 2 lines
You'd think that in a VStack of image and text, SwiftUI would support
some way to signal that you want the image to control the width of the
stack and the text to just flow after that. But I couldn't find a way,
so this CL just calculates the width of the icon and fixes the width
of the entire Destination view at that width.

Fixed: 1279876
Change-Id: I31d96baeb3220947b360ebbb8ca9900491f063b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3338528
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#951869}
NOKEYCHECK=True
GitOrigin-RevId: 48f99d5b8e29e40051d7d70318315c36c3623634