[iOS] Added functionality to perform download after receiving response
In the current Download API implementation for iOS 15+, we perform a
download when deciding a policy for navigation action. The issue with
this is that at this point in the download process, we don't have
access to the suggested filename and any other variables that we
receive at the point of receiving a response.

These changes will change the implementation so that we perform the
download when deciding a policy for navigation response if
|_shouldPerformDownload| is true. This will allow us to perform
a download after receiving a policy hint that a download
should be performed as well as fix the issue of not having access
to any response variables.

Bug: 1252380
Change-Id: I429c89ce2b3a1cdff75ebf960dcab676b9d0d97e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3336463
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#952357}
NOKEYCHECK=True
GitOrigin-RevId: af6580fc6235aa3f3a2a0b04fa308549fc0f456c