[ios] Remove MockWebState
The class web::FakeWebState supports settings the BrowserState
returned by GetBrowserState() so the sub-class is not needed.

The test was also incorrectly configuring the mock to return
a pointer to a deallocated object (as it was stored in a stack
allocated std::unique_ptr<T> that was destroyed by the end of
the SetUp() method).

Since the test do not require GetBrowserState() to return a
specific value anyway, just use a web::FakeWebState and do not
set the BrowserState.

Bug: none
Change-Id: Iad9db1661104c0bdeed52473bd24e7724d2c71e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3344757
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#952628}
NOKEYCHECK=True
GitOrigin-RevId: 834cb57f0ba1073ff3958540c48585bec26fa202