Move password generation requirements propagation to PasswordManager.
When server predictions are received, a call to
PasswordGenerationHelper::ProcessPasswordRequirements should be made.
Previously it was AutofillClient's responsibility to call this,
which lead to bugs, because there are multiple implementations of
AutofillClient (e.g. cross-platform, iOS, web_view), and it's easy to
forget this call.
This CL makes password generation requirements propagation a part of
PasswordManager::ProcessAutofillPredictions, and allows to have a
single point of calling it in the code.

With this CL ios/web_view will start receiving password generation
requirements.

Change-Id: I9d8a802971490690a4ba78ceb814b5fd81cc7200
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3311296
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#952927}
NOKEYCHECK=True
GitOrigin-RevId: ca9579b7a2f0aa0baafa96373ffefcca248c58f5