[ios][pandora2]🐻 Rollout to Beta
The experiment will start being effective in Beta when cherry-picked
in the Beta branch.

We also add an extra condition to not override the feature with the
trial when the feature is enabled with the about flag.

Launch bug: http://crbug.com/1199678

Bug: 1282200
Change-Id: I6fac68c2d4869fa1cfc747f8b4615acadb2c9d1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3353186
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#953867}
NOKEYCHECK=True
GitOrigin-RevId: 844653957a7cada1fab7a61b210b3a3b19937590