[iOS] Don't delay thumb strip appearance when message is appearing
The root cause of the bug is that the message is appearing at the same
time. TabGridViewController's -contentWillAppearAnimated: ties some
animations to the current transitionCoordinator if called with YES.
Here, the transitionCoordinator is the one for the message's appearance,
so the tab grid doesn't appear until that animation is done.

For the thumb strip, even though the thumb strip is appearing animated,
the animations are all handled via the ViewRevealingAnimatee protocol.

Bug: 1191154
Change-Id: I2b33c8c07d63027508dd0619e2bb1758f088539c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3303974
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#954969}
NOKEYCHECK=True
GitOrigin-RevId: 5718a70cb2aca0e72d75ab4c85c9ef770c6c8f32