[iOS] Add on-device encryption setting cells
One of the string is currently not translateable because we are not
yet entirely sure at what the final strings will be.  Indeed, the UI
https://docs.google.com/presentation/d/1Wd4O_6KMsuN4jLfsbdURdePcgJDFEVIXenDSSmZIlrI/edit?resourcekey=0-DwcN-hJHg8JgH08U-AMUWA#slide=id.gf660c2d721_2_67
can not be done as is, because we can’t just put a link in a cell.

Unittests and EGTests are to be done in a next CL. Do not close the
bug before.

Invite to opt-in screenshot on:
https://drive.google.com/file/d/1Yq1_D_W2V1nJwv43Ucy8J9YkfJvMFOCV/view?usp=sharing
and
Invite to learn-more for opted-in users screenshot on:
https://drive.google.com/file/d/1pqO6eDhA5nY3AVPB_TzIg4J8oXon3p-g/view?usp=sharing

Bug: 1202088
Change-Id: I500bf689f53353b1d2d67e80e8e673ae926f8cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3358248
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@google.com>
Cr-Commit-Position: refs/heads/main@{#955045}
NOKEYCHECK=True
GitOrigin-RevId: 5ce6fcacbde311b2ca5b0833b68ef9026e544d05