[ios] Keep WebState unrealized after favicon downloaded
Expose getter and setter to access the FaviconStatus for the
currently displayed page to WebState instead of having client
code access to the NavigationItem (which would force WebState
to become realized).

Change NavigationItem API to set the FaviconStatus. Instead
of exposing a non-const getter that return a reference to the
member field, expose a proper setter method. This make the
code a bit convoluted when the FaviconStatus is just partially
changed, but this make the code easier to understand.

Bug: 1264453
Change-Id: Idb8fd3ce84b4ef4e4a3bbf2fb4142690b53cb69e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3369964
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956506}
NOKEYCHECK=True
GitOrigin-RevId: 5e96365174115f42790066e8570238cd5415c0aa