[ios] Change WebStateListObserver to CheckedObserver
As part of investigating a crash where a WebStateListObserver is
destroyed without being removed from the observed WebStateList,
convert WebStateListObserver to CheckedObserver.

This allow to prevent WebStateList from causing crash when the
observers are notified after their destruction. Also add a CHECK
in WebStateListObserver to detect which observer is destroyed
but not unregistered. This will be converted to a DCHECK when
the error has been found and addressed.

Bug: 1283857
Change-Id: I3f0df81d17c44152e7a63f50a5623551958a7c2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372906
Reviewed-by: Louis Romero <lpromero@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956515}
NOKEYCHECK=True
GitOrigin-RevId: 4d9c6ca3d6bba6fe2ad38a194407450df99a31f6