[iOS] Move SettingsSwitchItem to table/cell
For the purpose of toggling media permissions in site info page, this CL
moves the SettingsSwitchItem and SettingsSwitchCell to table/cell, and
updates implementation and test references to this class.

SyncSwitchItem stays as is, but it uses TableViewSwitchCell instead of
SettingsSwitchCell. TableView Cell Catalog in Settings is also updated.

Bug: 1285140
Change-Id: I50a325c2f7c2b6759ebfc476f0e1104e65225537
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3371222
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956517}
NOKEYCHECK=True
GitOrigin-RevId: d816c108598623eeeeb65bb9a9e30f95c7ed2088