[ios] Move FontSizeJavaScriptFeature to //ios/chrome/web/font_size
This code is part of //ios/chrome/web/font_size so move it there.
It is used by the public provider implementation, which previously
required it to be part of the provider, but this is no longer the
case, so move it to its logical place.

Bug: 1201182
Change-Id: I400180cb04f14add9466a67dbe063b719d60455b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372790
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956550}
NOKEYCHECK=True
GitOrigin-RevId: dd5c83a9133d36c479deeb6b44db555243fc49c0