[iOS] Fix flake in SafeBrowsingServiceInitializationTest
This fixes flakiness that happens when tearing down V4Database.
V4Database::Destroy posts a task to destroy the database on
the database task runner. However, if WebTaskEnvironment
sometimes tears down that task runner before the task to destroy
the database has a chance to run. In that case, V4Database gets
destroyed at the point where the task is destroyed, on the
sequence that is tearing down the task runner, causing a DCHECK
since this is the wrong sequence.

To ensure that V4Database is destroyed before the database task
runner is torn down, this CL calls RunUntilIdle() on
WebTaskEnvironment to ensure that all pending tasks on the database
task runner get a chance to run.

Bug: 1285004
Change-Id: I44598268b69619ed2283accbe562b7ce6e358270
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3373608
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958152}
NOKEYCHECK=True
GitOrigin-RevId: 5d866ce675c94bdb975d6b37b4e5daf269a03bf0