[ios] Create ContentSuggestionsConsumer
This change replaces ContentSuggestionsDataSink/DataSource and
ContentSuggestionsCollectionUpdater with a consumer protocol that
sends the information needed from ContentSuggestionsMediator to
ContentSuggestionsViewController.

Method implementations are moved as needed as a part of this refactor,
but code functionality should not have changed.

Unused methods in related classes are removed.

Bug: 1285378
Change-Id: I43c114fb79625257151f76c464bb86cb6af18fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3345616
Reviewed-by: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958252}
NOKEYCHECK=True
GitOrigin-RevId: 8f7cebf3cd2205daf41d9c796a9b5613f10122f3