[iOS] Move logic for signing out the user to when they click no thanks
The current logic is to sign out the user when the advanced settings
finishes, but we want to do it only when the user tap on the secondary
action button ("No, Thanks"). This is to avoid having to sign-in again
if the user chooses to tap the primary action button ("Yes, I'm In").

Bug: 1282483
Change-Id: Ifec92447ed65846c0c42746dcee9e40868e3b250
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3355257
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Vincent Boisselle <vincb@google.com>
Commit-Queue: Veronique Nguyen <veronguyen@google.com>
Cr-Commit-Position: refs/heads/main@{#958378}
NOKEYCHECK=True
GitOrigin-RevId: 54dffc8777cbf9c938c183c41f351867978fec81