[ios] Expose `ui_utils` API as free functions
Add five free functions exposing the `ui_utils` API with the goal
to replace UI related methods of ChromeBrowserProvider class, but
do not use the new functions yet (as they have no implementation
in the internal repository yet).

This is a reland of https://crrev.com/c/3380978 fixed to not call
the provider API yet in ChromeBrowserProvider (as doing so broke
the internal repository).

This is part of a multi-CL change:
 => https://crrev.com/c/3381143
  - https://crrev.com/i/4447058
  - https://crrev.com/c/3383576
  - https://crrev.com/i/4451098
  - https://crrev.com/c/3380979

Bug: 1201182
Change-Id: I633d8d3b4c632f909425c24063345e4f26af8e35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3381143
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958478}
NOKEYCHECK=True
GitOrigin-RevId: a050c587481f0a5c44cb351c763cb98607905260