Update references of URLBlacklist and URLWhitelist in multiple files
Those two policeis are renamed to URLBlocklist and URLAllowlist.
The comments and test should use the new name as the new names are
preferred.
Also update the network annotation as the admin is encouraged to
migrate to the new policies as well.

Bug: 1283497
Change-Id: I0b85467811405323742ca81509667c62a826fc81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3362535
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958619}
NOKEYCHECK=True
GitOrigin-RevId: c7be31dd321bd0d97e4bc0d865804f82319f7f75