[ios] Use the new `app_utils` API
As the internal repository now provides implementation for the
`app_utils` API, convert code to use it instead of the deprecated
methods of ChromeBrowserProvider.

The methods are not removed as that would break the internal
repository (as the methods are still overridden and needs to
be until they are no longer called from public code).

This is part of a multi-CL change:
  - https://crrev.com/c/3383339
  - https://crrev.com/i/4447060
 => https://crrev.com/c/3383856
  - https://crrev.com/i/4451096
  - https://crrev.com/c/3380848

Bug: 1201182
Change-Id: I3c297b3bc2056212ff95fd643574cd00a3f75dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3383856
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958668}
NOKEYCHECK=True
GitOrigin-RevId: 559ce5f3ca1069ccab47be62604e27bcaeb9db97