[iOS][cpe-promo] Fix bug that image size is wrong on initial load
Currently, the size of the image is computed in -viewWillAppear:, using
the sizes of the root view of the view controller.

However, this size is not fully updated in -viewWillAppear:, because the
subviews aren't laid out yet, and the view size might be affected once
the subviews are laid out. This has caused the issue in the following
video shown:

https://drive.google.com/file/d/1-HbqBTgAY8RHe-0G1YvRRCqej6wRc0d1/view?usp=sharing&resourcekey=0-qBsGosBsAVztkjXOuUA4Rg

because when the back bounces back, the image size is re-calculated with
regards to the correct view size, with all other subviews laid out.

The same may potentially happen with all child implementation of
PromoStyleViewController.

Fixed: 1284654
Change-Id: Id73f909c195f4f000f2ff5f93d764bae77d6d865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3380024
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#958684}
NOKEYCHECK=True
GitOrigin-RevId: a650e111bfdcbccf04e63759187b5dd15f76c791