[iOS] Media permissions access in site information
In this CL, the page info coordinator hooks a newly implemented mediator
that interacts with the web state to get and set camera/microphone
access state with the page info view controller.

The mediator implements "PageInfoViewControllerPermissionsDelegate"
to provide information that is needed to display information in the view
controller's permissions section.

A new NSNumber category for Permission is also added to translate
web::Permission and web::PermissionState to NSNumber so that it could
be stored in an NSDictionary.

Bug: 1280626
Change-Id: I3bb815f3f7c4035ca913ca73e63e7ae43e3b15ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3344168
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#959169}
NOKEYCHECK=True
GitOrigin-RevId: 1f04e9a20963b156ed9795361f18c234f8753b75