[IOS]Add scrim view to the search mode.
When users tap on the search icon, a scrim view should appear over the
existing tabs waiting for the user to add text to the search bar.

I tried to use the search controller instead to make the process of
responding to the search query smoother but there were many complexities
due to trying to put the search controller inside a toolbar, so i ended
up using the regular search bar and its delegates.

Bug: 1286035
Change-Id: Ib9e2e993ce30f062674c3b2f6020d23bfbe382db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3379185
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#959248}
NOKEYCHECK=True
GitOrigin-RevId: b09701dc39bcd8edeb538ec67ffa7b9ec9100fac