[SH iOS] Change Main Frame DCHECK to return
We currently DCHECK to make sure the main frame is still available when
responding to the tap event that's supposed to remove the highlights
from the page. However, we're seeing crashes in stable when we try to
call the JS function, and there's a good chance that it's related to the
main frame (somehow) getting torn down before the event can be
processed. Thus, we're going to try an early return to avoid this crash.

Bug: 1287225
Change-Id: I41ad62be129fb4238d44ee23b088cabf8fc7ee5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3390209
Reviewed-by: Vincent Boisselle <vincb@google.com>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#959409}
NOKEYCHECK=True
GitOrigin-RevId: 53d4341322d7ddf080fec87d766f04a1f213ba7d